ai-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
numpy==2.3.2
Pillow==11.3.0
requests==2.32.4
